third_party/WebKit/LayoutTests/platform/chromium-win/fast/css3-text/css3-text-decoration/getComputedStyle/getComputedStyle-text-underline-position-expected.txt

Test to make sure -webkit-text-underline-position property returns values properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial value:
PASS e.style.getPropertyCSSValue('-webkit-text-underline-position') is null
FAIL computedStyle.webkitTextUnderlinePosition should be auto (of type string). Was undefined (of type undefined).
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSPrimitiveValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').cssText should be auto. Threw exception TypeError: Cannot read property 'cssText' of null

Value '':
PASS e.style.getPropertyCSSValue('-webkit-text-underline-position') is null
FAIL computedStyle.webkitTextUnderlinePosition should be auto (of type string). Was undefined (of type undefined).
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSPrimitiveValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').cssText should be auto. Threw exception TypeError: Cannot read property 'cssText' of null

Initial value (explicit):
PASS e.style.webkitTextUnderlinePosition is 'initial'
FAIL e.style.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL e.style.getPropertyCSSValue('-webkit-text-underline-position').cssText should be initial. Threw exception TypeError: Cannot read property 'cssText' of null
FAIL computedStyle.webkitTextUnderlinePosition should be auto (of type string). Was undefined (of type undefined).
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSPrimitiveValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').cssText should be auto. Threw exception TypeError: Cannot read property 'cssText' of null

Value 'auto':
PASS e.style.webkitTextUnderlinePosition is 'auto'
FAIL e.style.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSPrimitiveValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL e.style.getPropertyCSSValue('-webkit-text-underline-position').cssText should be auto. Threw exception TypeError: Cannot read property 'cssText' of null
FAIL computedStyle.webkitTextUnderlinePosition should be auto (of type string). Was undefined (of type undefined).
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSPrimitiveValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').cssText should be auto. Threw exception TypeError: Cannot read property 'cssText' of null

Value 'alphabetic':
PASS e.style.webkitTextUnderlinePosition is 'alphabetic'
FAIL e.style.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSPrimitiveValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL e.style.getPropertyCSSValue('-webkit-text-underline-position').cssText should be alphabetic. Threw exception TypeError: Cannot read property 'cssText' of null
FAIL computedStyle.webkitTextUnderlinePosition should be alphabetic (of type string). Was undefined (of type undefined).
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSPrimitiveValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').cssText should be alphabetic. Threw exception TypeError: Cannot read property 'cssText' of null

Value 'under':
PASS e.style.webkitTextUnderlinePosition is 'under'
FAIL e.style.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSPrimitiveValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL e.style.getPropertyCSSValue('-webkit-text-underline-position').cssText should be under. Threw exception TypeError: Cannot read property 'cssText' of null
FAIL computedStyle.webkitTextUnderlinePosition should be under (of type string). Was undefined (of type undefined).
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSPrimitiveValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').cssText should be under. Threw exception TypeError: Cannot read property 'cssText' of null

Ancestor inherits values from parent:
PASS e.style.getPropertyCSSValue('-webkit-text-underline-position') is null
FAIL computedStyle.webkitTextUnderlinePosition should be under (of type string). Was undefined (of type undefined).
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').toString() should be [object CSSPrimitiveValue]. Threw exception TypeError: Cannot call method 'toString' of null
FAIL computedStyle.getPropertyCSSValue('-webkit-text-underline-position').cssText should be under. Threw exception TypeError: Cannot read property 'cssText' of null

Value 'auto alphabetic':
PASS e.style.getPropertyCSSValue('-webkit-text-underline-position') is null

Value 'auto under':
PASS e.style.getPropertyCSSValue('-webkit-text-underline-position') is null

Value 'under under':
PASS e.style.getPropertyCSSValue('-webkit-text-underline-position') is null

Value 'under under under':
PASS e.style.getPropertyCSSValue('-webkit-text-underline-position') is null

PASS successfullyParsed is true

TEST COMPLETE
